Adds files
FormatRules.java contain the rules to change the text. You can revise, add or remove rules in this file. Every pair is a rule, the first fragment will be replaced by the second one.
Formatter.java and StrProcessor.java are necessary classes to built the GUI. Furthermore, these files contain the logic of the application.